+++ pkg/ociclient/upload.go
@@ -5,16 +5,14 @@ import (
 	"context"
 	"crypto/sha256"
 	"encoding/hex"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
 	"os"
-	"time"
+	"path/filepath"
 
-	digest "github.com/opencontainers/go-digest"
-	specs "github.com/opencontainers/image-spec/specs-go"
-	v1 "github.com/opencontainers/image-spec/specs-go/v1"
+	"github.com/k0kubun/go-ansi"
+	"github.com/schollz/progressbar/v3"
 )
 
 type Blob struct {
@@ -24,7 +22,7 @@ type Blob struct {
 	Permissions  os.FileMode
 }
 
-func uploadBlob(ctx context.Context, filePath, repoURL, jwtToken string) (*Blob, error) {
+func uploadBlob(ctx context.Context, filePath, repoURL, jwtToken string, showProgress bool, modelName string) (*Blob, error) {
 	file, err := os.Open(filePath)
 	if err != nil {
 		return nil, err
@@ -55,6 +53,32 @@ func uploadBlob(ctx context.Context, filePath, repoURL, jwtToken string) (*Blob,
 	chunkSize, err := determineChunkSize(filePath)
 	if err != nil {
 		return nil, fmt.Errorf("failed to determine chunk size: %w", err)
+	}
+
+	stat, err := file.Stat()
+	if err != nil {
+		return nil, err
+	}
+	size := stat.Size()
+	permissions := stat.Mode()
+
+	var bar *progressbar.ProgressBar
+	if showProgress {
+		// just the filename on the progress bar
+		filename := filepath.Base(filePath)
+		bar = progressbar.NewOptions(int(size),
+			progressbar.OptionSetWriter(ansi.NewAnsiStdout()), //you should install "github.com/k0kubun/go-ansi"
+			progressbar.OptionEnableColorCodes(true),
+			progressbar.OptionShowBytes(true),
+			progressbar.OptionSetWidth(15),
+			progressbar.OptionSetDescription(fmt.Sprintf("[cyan][%s][reset] Uploading %s...", modelName, filename)),
+			progressbar.OptionSetTheme(progressbar.Theme{
+				Saucer:        "[green]=[reset]",
+				SaucerHead:    "[green]>[reset]",
+				SaucerPadding: " ",
+				BarStart:      "[",
+				BarEnd:        "]",
+			}))
 	}
 
 	hasher := sha256.New()
@@ -98,16 +122,18 @@ func uploadBlob(ctx context.Context, filePath, repoURL, jwtToken string) (*Blob,
 		if uploadURL == "" {
 			return nil, fmt.Errorf("no upload URL returned")
 		}
+
+		if bar != nil {
+			bar.Add64(int64(n))
+		}
+	}
+
+	if bar != nil {
+		fmt.Println("")
+		bar.RenderBlank()
 	}
 
 	digest := fmt.Sprintf("sha256:%s", hex.EncodeToString(hasher.Sum(nil)))
-
-	stat, err := file.Stat()
-	if err != nil {
-		return nil, err
-	}
-	size := stat.Size()
-	permissions := stat.Mode()
 
 	req, err = http.NewRequest("PUT", uploadURL+"?digest="+digest, nil)
 	if err != nil {
@@ -131,104 +157,4 @@ func uploadBlob(ctx context.Context, filePath, repoURL, jwtToken string) (*Blob,
 		RelativePath: filePath,
 		Permissions:  permissions,
 	}, nil
-}
-
-func createAndUploadConfig(ctx context.Context, repoURL, jwtToken string, blobs []*Blob) (*Blob, error) {
-	now := time.Now()
-	config := v1.Image{
-		Created: &now,
-		Config: v1.ImageConfig{
-			Env:        []string{},
-			Entrypoint: []string{},
-			Cmd:        []string{},
-			Volumes:    map[string]struct{}{},
-			WorkingDir: "",
-			User:       "",
-		},
-		RootFS: v1.RootFS{
-			Type:    "layers",
-			DiffIDs: []digest.Digest{},
-		},
-		History: []v1.History{
-			{
-				Created:   &now,
-				CreatedBy: "file upload",
-			},
-		},
-	}
-
-	for _, blob := range blobs {
-		config.RootFS.DiffIDs = append(config.RootFS.DiffIDs, digest.Digest(blob.Digest))
-	}
-
-	configBytes, err := json.Marshal(config)
-	if err != nil {
-		return nil, err
-	}
-
-	tempFile, err := os.CreateTemp("", "config.json")
-	if err != nil {
-		return nil, err
-	}
-	defer os.Remove(tempFile.Name())
-
-	_, err = tempFile.Write(configBytes)
-	if err != nil {
-		return nil, err
-	}
-	tempFile.Close()
-
-	return uploadBlob(ctx, tempFile.Name(), repoURL, jwtToken)
-}
-
-func uploadManifest(ctx context.Context, blobs []*Blob, configBlob *Blob, repoURL, jwtToken, tag string) error {
-	var layers []v1.Descriptor
-	for _, blob := range blobs {
-		layers = append(layers, v1.Descriptor{
-			MediaType: "application/vnd.oci.image.layer.v1.tar+gzip",
-			Digest:    digest.Digest(blob.Digest),
-			Size:      blob.Size,
-			Annotations: map[string]string{
-				"org.opencontainers.image.layer.path":        blob.RelativePath,
-				"org.opencontainers.image.layer.permissions": fmt.Sprintf("%o", blob.Permissions),
-			},
-		})
-	}
-
-	manifest := v1.Manifest{
-		Versioned: specs.Versioned{
-			SchemaVersion: 2,
-		},
-		Config: v1.Descriptor{
-			MediaType: "application/vnd.oci.image.config.v1+json",
-			Digest:    digest.Digest(configBlob.Digest),
-			Size:      configBlob.Size,
-		},
-		Layers: layers,
-	}
-
-	manifestBytes, err := json.Marshal(manifest)
-	if err != nil {
-		return err
-	}
-
-	req, err := http.NewRequest("PUT", fmt.Sprintf("%s/manifests/%s", repoURL, tag), bytes.NewReader(manifestBytes))
-	if err != nil {
-		return err
-	}
-	req.Header.Set("Authorization", "Bearer "+jwtToken)
-	req.Header.Set("Content-Type", "application/vnd.oci.image.manifest.v1+json")
-	req.Header.Set("Content-Length", fmt.Sprintf("%d", len(manifestBytes)))
-
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		return err
-	}
-	defer resp.Body.Close()
-
-	if resp.StatusCode != http.StatusCreated && resp.StatusCode != http.StatusOK {
-		return fmt.Errorf("failed to upload manifest: %s", resp.Status)
-	}
-
-	return nil
 }
